reworked pokemonImageDownloader to have a stable progress bar and tag files with name+id

+++ model/PokemonImageDownloader.py
@@ -4,28 +4,43 @@ import os
 from tqdm import tqdm
 from urllib.parse import urlparse
 
-df = pd.read_csv('C:/Users/satvi/Desktop/Personal Projects/pokemon-price-checker-v2/model/dataset/pokemon-cards.csv')
+df = pd.read_csv('./dataset/pokemon-cards.csv')
 df = df.sample(frac=1).reset_index(drop=True)
 
-output_folder = 'pokemon_images'
+
+output_folder = '../dataset/pos_data'
 os.makedirs(output_folder, exist_ok=True)
 
 def download_images(df, image_column, output_folder):
-    for index, url in tqdm(enumerate(df[image_column]), total=len(df)):
-        try:    
-            name = df.loc[index, 'name']
-            filename = f"{name.replace(' ', '_')}.jpg"
-            image_path = os.path.join(output_folder, filename)
+    with tqdm(total=len(df), ncols= 80, desc="Downloading images") as pbar:
+        errors = 0
+        for index, url in enumerate(df[image_column]):
+            try:    
+                name = df.loc[index, 'name'] + '_' + df.loc[index, 'id']
+                filename = f"{name.replace(' ', '_')}.jpg"
+                filename = filename.replace('"', '')
+                filename = filename.replace('δ', 'delta')
 
-            response = requests.get(url, timeout=10)
-            response.raise_for_status()
+                #check if filename already exists in output_folder
+                if os.path.exists(os.path.join(output_folder, filename)):
+                    tqdm.write(f"Already downloaded: {filename}")
+                else:
+                    image_path = os.path.join(output_folder, filename)
 
-            with open(image_path, 'wb') as f:
-                f.write(response.content)
+                    response = requests.get(url, timeout=10)
+                    response.raise_for_status()
 
-            print(f"\n Saved: {image_path}")
+                    with open(image_path, 'wb') as f:
+                        f.write(response.content)
 
-        except Exception as e:
-            print(f"\n Failed to download {url}: {e}")
+            except Exception as e:
+                # add number of errors to a counter
+                errors += 1
+                tqdm.write(f"Failed to download {url}: {e}")
+
+            if(index%100 == 0):
+                tqdm.write(f"Downloaded {index} images with {errors} errors")
+            
+            pbar.update(1)
 
 download_images(df, 'image_url', output_folder)
